requirements.txt
Django
Pillow
django-ckeditor
django-constance[database]